refactor(Simulations): split objects
